Move balance graph to disclosure in address page
